render data from table
This is done by creating a 'global' variable `window.rowValues`,
that can be accessed from both uploader.js and chart.js.
When loading a file, the row values are stored in it, so they
can be accessed when the call to chart.js is made.

There, each row is made into an object ('dictionary') so it can
be parsed by the rendering code.